+++ src/components/Common/index.tsx
@@ -1,0 +1,10 @@
+export { default as Carousel } from './Carousel';
+export { default as Category } from './Category';
+export { default as CategoryButton } from './CategoryButton';
+export { default as CustomButton } from './CustomButton';
+export { default as CustomInput } from './CustomInput';
+export { default as Header } from './Header';
+export { default as Layout } from './Layout';
+export { default as MainPageHeader } from './MainPageHeader';
+export { default as TapBar } from './TapBar';
+export { default as Review } from './Review';
