--- src/components/Post/PostComponent.tsx
@@ -1,5 +1,0 @@
-const PostComponent = () => {
-  return <div>PostComponent</div>;
-};
-
-export default PostComponent;
