+++ src/pages/Auth/SocialLogin/KaKao/SocialKakaoButton/index.tsx
@@ -1,5 +1,5 @@
-import kakao_login_medium_narrow from '../../../../assets/images/kakao_login_medium_narrow.webp';
-import config from '../../../../config';
+import kakao_login_medium_narrow from '../../../../../assets/images/kakao_login_medium_narrow.webp';
+import config from '../../../../../config';
 
 const SocialKakaoButton = () => {
   const kakaoURL = `https://kauth.kakao.com/oauth/authorize?client_id=${config.KAKAO_REST_API_KEY}&redirect_uri=${config.KAKAO_REDIRECT_URL}&response_type=code`;
