+++ src/pages/Auth/index.tsx
@@ -1,0 +1,4 @@
+export { default as Login } from './Login';
+export { default as Register } from './Register';
+export { default as SocialKakaoButton } from './SocialLogin/KaKao/SocialKakaoButton';
+export { default as SocialKakaoRedirectPage } from './SocialLogin/KaKao/SocialKakaoRedirectPage';
