+++ src/pages/HomePage/index.tsx
@@ -1,8 +1,8 @@
-import Layout from '../components/Common/Layout';
-import MainPageHeader from '../components/Common/MainPageHeader';
-import TapBar from '../components/Common/TapBar';
-import Carousel from '../components/Common/Carousel';
-import Category from '../components/Common/Category';
+import Layout from '../../components/Common/Layout';
+import MainPageHeader from '../../components/Common/MainPageHeader';
+import TapBar from '../../components/Common/TapBar';
+import Carousel from '../../components/Common/Carousel';
+import Category from '../../components/Common/Category';
 import { InputGroup, InputLeftElement, Input } from '@chakra-ui/react';
 import { SearchIcon } from '@chakra-ui/icons';
