+++ src/pages/Map/index.tsx
@@ -1,0 +1,5 @@
+export { default as CustomMapMarker } from './CustomMapMarker';
+export { default as ListComponent } from './ListComponent';
+export { default as Locations } from './Locations';
+export { default as MapPage } from './MapPage';
+// export { default as ValleyClusterer } from './ValleyClusterer';
