+++ src/pages/MyPage/index.tsx
@@ -9,15 +9,15 @@ import {
   Divider,
 } from '@chakra-ui/react';
 // import { useUserStore } from '주소'; //여기다가 유저데이터 store 넣으면 됨.
-import Layout from '../components/Common/Layout';
-import InstaImage from '../components/Common/Image/InstaImage';
-import { fetchReviews } from '../api/ReviewsApi/ReviewsApi';
+import Layout from '../../components/Common/Layout';
+import InstaImage from '../../components/Common/Image/InstaImage';
+import { fetchReviews } from '../../api/ReviewsApi/ReviewsApi';
 import { Link } from 'react-router-dom';
-import { ReviewType } from '../types/ReviewType/ReviewType';
-import ProfileImage from '../components/Common/Image/ProfileImage';
-import valley1 from '../../valley1.png';
-import TapBar from '../components/Common/TapBar';
-import Header from '../components/Common/Header';
+import { ReviewType } from '../../types/ReviewType/ReviewType';
+import ProfileImage from '../../components/Common/Image/ProfileImage';
+import valley1 from '../../../valley1.png';
+import TapBar from '../../components/Common/TapBar';
+import Header from '../../components/Common/Header';
 
 const MyPage: React.FC = () => {
   const reviews = useLoaderData() as ReviewType[];
