+++ src/routes/index.tsx
@@ -1,1 +1,1 @@
-export {};
+export { default as PrivateRoute } from './PrivateRoute';
