--- src/store.ts
@@ -1,1 +1,0 @@
-export {};
