+++ src/types/index.ts
@@ -1,2 +1,2 @@
 export * from './ReviewType/ReviewType';
-export * from './ListType/types';
+// export * from './ListType/types';
